# Conflicts:
#	src/ClientPackage/View/GUIResources/Class/MatchController.java
#	src/ClientPackage/View/GUIResources/Class/ShopController.java
#	src/ClientPackage/View/GUIResources/customComponent/CouncilorHandler.java
#	src/ClientPackage/View/GUIResources/customComponent/PermitCardHandler.java
#	src/Utilities/Class/Constants.java

+++ src/Utilities/Class/CircularArrayList.java
@@ -12,7 +12,7 @@ public class CircularArrayList<E> extends ArrayList<E> {
     @Override
     public E get(int index) {
         if (index < 0)
-            index = index + ( (int) abs(index/size()) + 1)*size();
+            index = index + ( abs(index/size()) + 1)*size();
         if (index > 0)
             index = index%size();
         return super.get(index);
